requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.11
click==8.0.3
deepgram-sdk==0.2.4
fastapi==0.73.0
frozenlist==1.3.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
multidict==6.0.2
pydantic==1.9.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.0.1
uvicorn==0.17.4
websockets==10.1
yarl==1.7.2
